requirements.txt
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
graphql-core==3.2.3
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
libcst==1.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
pydantic==1.10.13
Pygments==2.17.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
rich==13.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlmodel==0.0.12
starlette==0.27.0
strawberry-graphql==0.215.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
